Rename DBFilters to DBTools, as if does not only filter anymore
